fix test_mem

+++ TP2-SO-2025/Userland/Shell/commands/test_mm.c
@@ -54,6 +54,7 @@ uint64_t _test_mm(uint64_t argc, char *argv[]) {
       if (mm_rqs[i].address)
         freeMemory(mm_rqs[i].address);
     // Report progress to the user after each iteration
-    printf("test_mm: iteration %d completed; requested_blocks=%d total_bytes=%u\n", count, rq, total);
+    /* Use %d because the project's printf implementation supports %d (not necessarily %u) */
+    printf("test_mm: iteration %d completed; requested_blocks=%d total_bytes=%d\n", count, rq, (int)total);
   }
 }
